requirements.txt
streamlit
os-sys
pandas
datetime
razorpay
db-sqlite3
googletrans-py